Azure/LAB001/requirements.txt
streamlit 
azure-storage-blob 
pymssql
python-dotenv